refactor(cli): split commands into modules
